Number of people buying multiple energy drinks, and number of those people who prefer citrus

+++ Sp2025 CSC 221 Project 2.cpp
@@ -16,7 +16,7 @@ int main()
     totalCustServey = 16500;
     perMultDrink = 0.15;
     perCitrusPref = 0.58;
-    numCustMultDrink = totalCustServey * perMultDrink;
+    numCustMultDrink = int(totalCustServey * perMultDrink);
     numCitrusPref = int(numCustMultDrink * perCitrusPref);
